Move mobile banner to slot fill
