+++ client/layout/index.js
@@ -29,8 +29,9 @@ import { Controller, getPages } from './controller';
 import { Header } from '../header';
 import Notices from './notices';
 import TransientNotices from './transient-notices';
-import './navigation';
 import '~/activity-panel';
+import '~/mobile-banner';
+import './navigation';
 
 const StoreAlerts = lazy( () =>
 	import( /* webpackChunkName: "store-alerts" */ './store-alerts' )
