+++ client/mobile-banner/banner.scss
@@ -4,15 +4,18 @@ $banner-height: 56px;
 // This class is a hack, added conditionally when displaying the app banner, it
 // can be refactored away when the activity panel goes away.
 .woocommerce-layout__show-app-banner {
-	padding-top: $banner-height;
-
-	@include breakpoint( '>782px' ) {
-		padding-top: 0;
+	@include breakpoint( '<782px' ) {
+		.woocommerce-layout__header-wrapper {
+			padding-top: $banner-height;
+		}
 	}
 }
 
 .woocommerce-mobile-app-banner {
 	background-color: $woocommerce-brand-purple;
+	position: fixed;
+	top: 0;
+	left: 0;
 	width: 100%;
 	display: flex;
 	height: $banner-height;
